// src/app/light-commissioning/shared/model/stats/profilepush-node-stats.model.ts
export interface DeviceStats {
  commissioned: number;
  discovered: number;
  up: number;
  down: number;
  normal: number;
  enocean: number;
  sink: number;
  ledController: number;
  relayController: number;
  diController: number;
  mokoBand: number;
  total: number;
  autoModeEnable: number;
  autoModeDisable: number;
}